add API token settings component with update functionality
